Function inanis::engine::qsearch::assign_move_scores

fn assign_move_scores(
    context: &SearchContext<'_>, 
    moves: &[Move], 
    move_scores: &mut [i16], 
    moves_count: usize
)

Assigns scores for moves by filling move_scores array with moves_count length, based on current context. Move ordering in quiescence search is mainly based on SEE and works as follows:

for every en passant, assign 0

for every capture with promotion, assign value of the promoted piece

for rest of the moves, assign SEE result