Function inanis::engine::search::run

pub fn run<const PV: bool>(
    context: &mut SearchContext<'_>, 
    depth: i8, 
    ply: u16, 
    alpha: i16, 
    beta: i16, 
    allow_null_move: bool, 
    friendly_king_checked: bool
) -> i16

Entry point of the regular search, with generic PV parameter determining if the current node is a PV (principal variation) node in PVS framework. The implementation contains a typical alpha-beta approach, together with a bunch of reduction and prunings to optimize search. The most important parameter here, context contains the current state of the search, board state, statistics, and is passed by reference to all nodes. Besides obvious parameters like depth, ply, alpha and beta, there’s also allow_null_move which prevents two null move checks in a row, and friendly_king_checked which is used to share friendly king check status between nodes (it’s always calculated one depth earlier, as it’s used as one of the LMR constraints).

Search steps for PV node:

Test of initial constraints: abort flag, forced depth

Test if the enemy king is checked

Test if there’s threefold repetition draw, fifty move rule draw or insufficient material draw

Switch to quiescence search if the depth is equal to zero

Read from transposition table, return score if possible or update alpha/beta (https://www.chessprogramming.org/Transposition_Table)

Generate evasion mask if the friendly king is checked

Main loop:

Late move reduction (https://www.chessprogramming.org/Late_Move_Reductions)

Reduction pruning (https://www.chessprogramming.org/History_Leaf_Pruning)

PVS framework (https://www.chessprogramming.org/Principal_Variation_Search)

Test of abort flag

Test if stalemate draw is detected

Update transposition table

Search steps for non-PV node:

Test of initial constraints: abort flag, forced depth, max nodes count

Test if the enemy king is checked

Test if there’s threefold repetition draw, fifty move rule draw or insufficient material draw

Switch to quiescence search if the depth is equal to zero

Read from transposition table, return score if possible or update alpha/beta (https://www.chessprogramming.org/Transposition_Table)

Razoring (https://www.chessprogramming.org/Razoring)

Static null move pruning (https://www.chessprogramming.org/Reverse_Futility_Pruning)

Null move pruning (https://www.chessprogramming.org/Null_Move_Pruning)

Generate evasion mask if the friendly king is checked

Main loop:

Late move pruning (https://www.chessprogramming.org/Futility_Pruning#MoveCountBasedPruning)

Late move reduction (https://www.chessprogramming.org/Late_Move_Reductions)

Reduction pruning (https://www.chessprogramming.org/History_Leaf_Pruning)

PVS framework (https://www.chessprogramming.org/Principal_Variation_Search)

Test of abort flag

Test if stalemate draw is detected

Update transposition table