Module inanis::evaluation::pst

Re-exports

pub use bishop::BISHOP_PST_PATTERN;

pub use king::KING_PST_PATTERN;

pub use knight::KNIGHT_PST_PATTERN;

pub use pawn::PAWN_PST_PATTERN;

pub use queen::QUEEN_PST_PATTERN;

pub use rook::ROOK_PST_PATTERN;

Modules

bishop

king

knight

pawn

queen

rook

Constants

KING_BUCKETS

KING_BUCKETS_COUNT

Functions

evaluate

Evaluates piece-square table value on the board and returns score from the white color perspective (more than 0 when advantage, less than 0 when disadvantage).

get_pst_value

Gets a PST value for the specified piece, pov, king_square and square (relative perspective).

recalculate_incremental_values

Recalculates incremental counters on the board. This function should be called only if really necessary, as it’s too slow in regular search.