Function inanis::evaluation::pst::evaluate

pub fn evaluate(board: &Board) -> EvaluationResult

Evaluates piece-square table value on the board and returns score from the white color perspective (more than 0 when advantage, less than 0 when disadvantage). This evaluator sums all values of the pieces for the specified squares, using incremental counters in board.