Function inanis::interface::terminal::handle_qperft

fn handle_qperft(input: Vec<&str>)

Handles qperft [depth] [threads_count] [hashtable_size_mb], qperft [depth] [threads_count] [hashtable_size_mb] fen [fen] and qperft [depth] [threads_count] [hashtable_size_mb] moves [moves] commands by running a quick perft test to the depth specified by depth parameter. This kind of perft also supports multithreading (specified by threads_count) and caching results in the hashtable (with size specified by hashtable_size_mb). The initial position can be specified by FEN, a list of moves, or just omitted (so the default start position will be taken).