Function inanis::interface::uci::handle_go

fn handle_go(parameters: &[String], state: Arc<Mutex<UciState>>)

Handles go [parameters] command by running a new search for a position which was set using position command. Supported parameters:

wtime x - amount of total time for white in milliseconds

btime x - amount of total time for black in milliseconds

winc x - incremental time for white

binc x - incremental time for black

depth x - fixed depth, where the search will stop

nodes x - fixed nodes count, after which the search will try to stop as soon as possible

movetime x - fixed time allocated for the search in milliseconds

movestogo x - amount of moves, after which the time will be increased

infinite - tells the search to run until it reaches the maximal depth for the engine

searchmoves [moves] - restricts search to the provided moves list

ponder - tells the search to run in the ponder mode (thinking on the opponent’s time)