Module inanis::state::movegen

Structs

MagicSquare

Constants

BISHOP_MAGIC_NUMBERS 🔒

BISHOP_SHIFTS 🔒

ROOK_MAGIC_NUMBERS 🔒

ROOK_SHIFTS 🔒

Statics

BISHOP_SQUARES 🔒

ROOK_SQUARES 🔒

Functions

apply_bishop_magic 🔒

Applies bishop magic for the square specified by square, using built-in magic number from BISHOP_MAGIC_NUMBERS.

apply_magic_for_square 🔒

Applies a magic number for a set of permutations, attacks and square.

apply_rook_magic 🔒

Applies rook magic for the square specified by square, using built-in magic number from ROOK_MAGIC_NUMBERS.

generate_bishop_magic_number

Generates a bishop magic number for the square specified by square.

generate_magic_number 🔒

Generates a magic number for a set of permutations and attacks, using shift proper for the specified square.

generate_rook_magic_number

Generates a rook magic number for the square specified by square.

get_attacks 🔒

Helper function to get all possible to move squares, considering occupancy_bb, starting from the square specified by square and going into the direction.

get_bishop_attacks 🔒

Gets a bishop attacks for the square specified by square, occupancy occupancy_bb.

get_bishop_mask 🔒

Gets a bishop mask for the square specified by square, without considering occupancy.

get_bishop_moves

Gets a bishop moves for the square specified by square, considering occupancy_bb.

get_king_moves

Gets a king moves for the square specified by square, without considering an occupancy.

get_knight_moves

Gets a knight moves for the square specified by square, without considering an occupancy.

get_permutation 🔒

Gets index-th permutation of the mask.

get_queen_moves

Gets a queen moves for the square specified by square, considering occupancy_bb.

get_rook_attacks 🔒

Gets a rook attacks for the square specified by square, considering occupancy_bb.

get_rook_mask 🔒

Gets a rook mask for the square specified by square, without considering occupancy.

get_rook_moves

Gets a rook moves for the square specified by square, considering occupancy_bb.

init

Initialize movegen for rooks and bishops.