Module inanis::state::movescan

Structs

Move

MoveFlags

Functions

get_piece_mobility

Gets PIECE mobility (by counting all possible moves at the position specified by board) with color and increases dangered_king_fields if the enemy king is near to the fields included in the mobility.

scan_pawn_moves

Generates all possible non-captures (if CAPTURES is false) or all possible captures (if CAPTURES is true) for the pawns at the position specified by board, stores them into moves list (starting from index) and returns index of the first free slot. Use evasion_mask with value different than u64::MAX to restrict generator to the specified fields (useful during checks).

scan_pawn_moves_diagonal_attacks 🔒

Generates all possible captures for the pawns toward the direction specified by DIR and at the position specified by board, stores them into moves list (starting from index) and returns index of the first free slot. Use evasion_mask with value different than u64::MAX to restrict generator to the specified fields (useful during checks).

scan_pawn_moves_double_push 🔒

Generates all possible double pushes for the pawns at the position specified by board, stores them into moves list (starting from index) and returns index of the first free slot. Use evasion_mask with value different than u64::MAX to restrict generator to the specified fields (useful during checks).

scan_pawn_moves_single_push 🔒

Generates all possible single pushes for the pawns at the position specified by board, stores them into moves list (starting from index) and returns index of the first free slot. Use evasion_mask with value different than u64::MAX to restrict generator to the specified fields (useful during checks).

scan_piece_moves

Generates all possible non-captures (if CAPTURES is false) or all possible captures (if CAPTURES is true) for the PIECE at the position specified by board, stores them into moves list (starting from index) and returns index of the first free slot. Use evasion_mask with value different than u64::MAX to restrict generator to the specified fields (useful during checks).