Function inanis::state::movescan::scan_pawn_moves_double_push

fn scan_pawn_moves_double_push(
    board: &Board,
    moves: &mut Moves,
    index: usize,
    evasion_mask: u64
) -> usize

Generates all possible double pushes for the pawns at the position specified by board, stores them into moves list (starting from index) and returns index of the first free slot. Use evasion_mask with value different than u64::MAX to restrict generator to the specified squares (useful during checks).