Struct inanis::state::movescan::Move

pub struct Move {
    pub data: u16,
}

Fields
data: u16
Implementations

impl Move

pub fn new(from: usize, to: usize, flags: u8) -> Self

Constructs a new instance of Move with stored from, to and flags.

pub const fn new_from_raw(data: u16) -> Self

Constructs a new instance of Move using raw bits, which will be directly used as a data.

pub fn new_random() -> Self

Constructs a new instance of Move with random values, not restricted by chess rules.

pub fn get_from(&self) -> usize

Gets source square from the internal data.

pub fn get_to(&self) -> usize

Gets destination square from the internal data.

pub fn get_flags(&self) -> u8

Gets flags from the internal data.

pub fn get_promotion_piece(&self) -> usize

Gets promotion piece based on the flags saved in the internal data.

pub fn is_some(&self) -> bool

Checks if there is any data present or it’s just a default move.

pub fn is_empty(&self) -> bool

Checks if there is any data present or it’s just a default move.

pub fn is_single_push(&self) -> bool

Checks if the move is a single push.

pub fn is_double_push(&self) -> bool

Checks if the move is a double push.

pub fn is_quiet(&self) -> bool

Checks if the move is quiet (single or double pushes).

pub fn is_capture(&self) -> bool

Checks if the move is a capture (excluding en passant, but including promotions).

pub fn is_en_passant(&self) -> bool

Checks if the move is en passant.

pub fn is_promotion(&self) -> bool

Checks if the move is a promotion (including captures).

pub fn is_castling(&self) -> bool

Checks if the move is a short or long castling.

pub fn is_legal(&self, board: &Board) -> bool

Checks if the move is legal, using board as the context.

impl Move

pub fn from_short_notation( text: &str, board: &mut Board ) -> Result<Move, String>

Converts short-notated move (e4, Rc8, Qxb6) in text into the Move instance, using the board as context. Returns Err with the proper message if text couldn’t be parsed correctly.

pub fn from_long_notation(text: &str, board: &Board) -> Result<Move, String>

Converts long-notated move (e2e4, a1a8) in text into the Move instance, using the board as context. Returns Err with the proper message if text couldn’t be parsed correctly.

pub fn to_long_notation(self) -> String

Converts move into the long notation (e2e4, a1a8).