Module inanis::state::patterns

Constants

BOX_PATTERNS

DIAGONAL_PATTERNS

FRONT_PATTERNS

JUMP_PATTERNS

RAIL_PATTERNS

Functions

generate_boxes

Generates box patterns for all squares.

generate_diagonals

Generates diagonal patterns for all squares.

generate_fronts

Generates front patterns for all squares.

generate_jumps

Generates jump patterns for all fields.

generate_rails

Generates rail patterns for all squares.

get_box

Get a box pattern for the square specified by square.

get_diagonals

Gets a diagonal pattern for the square specified by square_.

get_file

Gets a file pattern for the square specified by square.

get_front

Get a front pattern for the square specified by square, from the color perspective.

get_jumps

Get a jumps pattern for the square specified by square.

get_rail

Get a rail pattern for the square specified by file.

get_rank

Gets a rank pattern for the square specified by square.