Struct inanis::state::patterns::PatternsContainer

pub struct PatternsContainer {
    pub file_patterns: [u64; 64],
    pub rank_patterns: [u64; 64],
    pub diagonal_patterns: [u64; 64],
    pub jump_patterns: [u64; 64],
    pub box_patterns: [u64; 64],
    pub rail_patterns: [u64; 8],
    pub star_patterns: [u64; 64],
    pub front_patterns: [[u64; 64]; 2],
}

Fields
file_patterns: [u64; 64]rank_patterns: [u64; 64]diagonal_patterns: [u64; 64]jump_patterns: [u64; 64]box_patterns: [u64; 64]rail_patterns: [u64; 8]star_patterns: [u64; 64]front_patterns: [[u64; 64]; 2]
Implementations

impl PatternsContainer

pub fn get_file(&self, field_index: usize) -> u64

Gets a file pattern for the field specified by field_index.

. . . x . . . .
. . . x . . . .
. . . x . . . .
. . . x . . . .
. . . o . . . .
. . . x . . . .
. . . x . . . .
. . . x . . . .

pub fn get_rank(&self, field_index: usize) -> u64

Gets a rank pattern for the field specified by field_index.

. . . . . . . .
. . . . . . . .
. . . . . . . .
. . . . . . . .
x x x o x x x x
. . . . . . . .
. . . . . . . .
. . . . . . . .

pub fn get_diagonals(&self, field_index: usize) -> u64

Gets a diagonal pattern for the field specified by field_index.

. . . . . . . x
x . . . . . x .
. x . . . x . .
. . x . x . . .
. . . o . . . .
. . x . x . . .
. x . . . x . .
x . . . . . x .

pub fn get_jumps(&self, field_index: usize) -> u64

Get a jumps pattern for the field specified by field_index.

. . . . . . . .
. . . . . . . .
. . x . x . . .
. x . . . x . .
. . . o . . . .
. x . . . x . .
. . x . x . . .
. . . . . . . .

pub fn get_box(&self, field_index: usize) -> u64

Get a box pattern for the field specified by field_index.

. . . . . . . .
. . . . . . . .
. . . . . . . .
. . x x x . . .
. . x o x . . .
. . x x x . . .
. . . . . . . .
. . . . . . . .

pub fn get_rail(&self, file: usize) -> u64

Get a rail pattern for the field specified by field_index.

. . x . x . . .
. . x . x . . .
. . x . x . . .
. . x . x . . .
. . x o x . . .
. . x . x . . .
. . x . x . . .
. . x . x . . .

pub fn get_star(&self, field_index: usize) -> u64

Get a star pattern for the field specified by field_index.

. . . . . . . .
. . . . . . . .
. . . . . . . .
. . x . x . . .
. . . o . . . .
. . x . x . . .
. . . . . . . .
. . . . . . . .

pub fn get_front(&self, color: usize, field_index: usize) -> u64

Get a front pattern for the field specified by field_index, from the color perspective.

. . x x x . . .
. . x x x . . .
. . x x x . . .
. . x x x . . .
. . . o . . . .
. . . . . . . .
. . . . . . . .
. . . . . . . .

pub fn regenerate_files(&mut self)

Generates file patterns for all fields.

pub fn regenerate_ranks(&mut self)

Generates rank patterns for all fields.

pub fn regenerate_diagonals(&mut self)

Generates diagonal patterns for all fields.

pub fn regenerate_jumps(&mut self)

Generates jump patterns for all fiellds.

pub fn regenerate_boxes(&mut self)

Generates box patterns for all fields.

pub fn regenerate_rails(&mut self)

Generates rail patterns for all fields.

pub fn regenerate_stars(&mut self)

Generates star patterns for all fields.

pub fn regenerate_fronts(&mut self)

Generates front patterns for all fields.