Function inanis::state::text::fen::epd_to_board

pub fn epd_to_board(
    epd: &str,
    evaluation_parameters: Option<Arc<EvaluationParameters>>,
    zobrist_container: Option<Arc<ZobristContainer>>,
    patterns_container: Option<Arc<PatternsContainer>>,
    see_container: Option<Arc<SEEContainer>>,
    magic_container: Option<Arc<MagicContainer>>
) -> Result<ParsedEPD, String>

Converts epd into the Board, using provided containers. If the parameter is None, then the new container is created. Returns Err with proper error message if epd couldn’t be parsed correctly.