Function inanis::state::text::fen::fen_to_board

pub fn fen_to_board(
    fen: &str,
    evaluation_parameters: Option<Arc<EvaluationParameters>>,
    zobrist_container: Option<Arc<ZobristContainer>>,
    patterns_container: Option<Arc<PatternsContainer>>,
    see_container: Option<Arc<SEEContainer>>,
    magic_container: Option<Arc<MagicContainer>>
) -> Result<Board, String>

Converts fen into the Board, using provided containers. If the parameter is None, then the new container is created. Returns Err with proper error message if fen couldn’t be parsed correctly.