Module inanis::state::zobrist

Statics

ACTIVE_COLOR_HASH

CASTLING_HASHES

EN_PASSANT_HASHES

PIECE_HASHES

Functions

get_active_color_hash

Gets active color hash.

get_castling_right_hash

Gets castling right hash based on the current ones and the desired change specified by right.

get_en_passant_hash

Gets en passant hash for the file.

get_piece_hash

Gets piece hash with the color for the field specified by field_index.

init

Initializes Zobrist hashes by filling arrays with random u64 values.

recalculate_hash

Recalculates Zobrist hash for the board.

recalculate_pawn_hash

Recalculates pawn hash for the board.