Module inanis::utils::test

Structs

TestContext 🔒

TestPosition 🔒

Functions

load_positions 🔒

Loads positions from the epd_filename and parses them into a list of TestPosition. Returns Err with a proper error message if the file couldn’t be parsed.

run

Runs a test by performing a fixed-depth search for the positions loaded from the epd_filename file, using hashtable with size specified in transposition_table_size. To classify the test as successful, the last iteration has to return the correct best move.

run_internal 🔒

Internal test function, called by run.