Crate inanis

The main page of the Inanis documentation. Feel free to explore it by going into the specific module below, or by clicking “See all inanis’s items” on the left panel to see every possible item.

Homepage: https://github.com/Tearth/Inanis

Modules

cache

engine

evaluation

interface

perft

state

tablebases

testing

tuning

utils