Function inanis::engine::search::get_next_move

fn get_next_move(
    context: &mut SearchContext<'_>, 
    stage: &mut MoveGeneratorStage, 
    moves: &mut [Move], 
    move_scores: &mut [i16], 
    move_index: &mut usize, 
    moves_count: &mut usize, 
    hash_move: Move, 
    evasion_mask: u64, 
    ply: u16
) -> Option<Move>

Gets the next move to analyze. This function acts as pseudo-iterator and takes care about managing move generator stages, which is basically a state machine (https://en.wikipedia.org/wiki/Finite-state_machine) with following rules:

MoveGeneratorStage::ReadyToCheckHashMove - default state, returns transposition table move if possible

MoveGeneratorStage::ReadyToGenerateCaptures - generates all captures in the position

MoveGeneratorStage::Captures - returns subsequent elements until the end or score is less than MOVE_ORDERING_WINNING_CAPTURES_OFFSET

MoveGeneratorStage::ReadyToGenerateQuietMoves - generates all quiet moves in the position

MoveGeneratorStage::AllGenerated - returns subsequent elements until the end

Both MoveGeneratorStage::ReadyToGenerateCaptures and MoveGeneratorStage::ReadyToGenerateQuietMoves are generating moves and assigning scores for move ordering purposes. If the last stage is set and there are no more moves, None is returned.