Module inanis::state::movegen

Structs

MagicField

Statics

BISHOP_FIELDS

BISHOP_MAGIC_NUMBERS

BISHOP_SHIFTS

ROOK_FIELDS

ROOK_MAGIC_NUMBERS

ROOK_SHIFTS

Functions

apply_bishop_magic_for_field

Applies bishop magic for the field specified by field_index, using built-in magic number from BISHOP_MAGIC_NUMBERS.

apply_magic_for_field

Applies a magic number for a set of permutations, attacks and field.

apply_rook_magic_for_field

Applies rook magic for the field specified by field_index, using built-in magic number from ROOK_MAGIC_NUMBERS.

generate_bishop_number_for_field

Generates a bishop magic number for the field specified by field_index.

generate_magic_number

Generates a magic number for a set of permutations and attacks, using shift proper for the specified field.

generate_rook_number_for_field

Generates a rook magic number for the field specified by field_index.

get_attacks

Helper function to get all possible to move fields, considering occupancy, starting from the field specified by field_index and going into the direction.

get_bishop_attacks

Gets a bishop attacks for the field specified by field_index, occupancy occupancy.

get_bishop_mask

Gets a bishop mask for the field specified by field_index, without considering occupancy.

get_bishop_moves

Gets a bishop moves for the field specified by field_index, considering occupancy.

get_king_moves

Gets a king moves for the field specified by field_index, without considering an occupancy.

get_knight_moves

Gets a knight moves for the field specified by field_index, without considering an occupancy.

get_permutation

Gets index-th permutation of the mask.

get_queen_moves

Gets a queen moves for the field specified by field_index, considering occupancy.

get_rook_attacks

Gets a rook attacks for the field specified by field_index, considering occupancy.

get_rook_mask

Gets a rook mask for the field specified by field_index, without considering occupancy.

get_rook_moves

Gets a rook moves for the field specified by field_index, considering occupancy.

init

Initializes move generator by applying built-in magic numbers (ROOK_MAGIC_NUMBERS and BISHOP_MAGIC_NUMBERS).