Module inanis::state::patterns

Statics

BOX_PATTERNS

DIAGONAL_PATTERNS

FILE_PATTERNS

FRONT_PATTERNS

JUMP_PATTERNS

RAIL_PATTERNS

RANK_PATTERNS

STAR_PATTERNS

Functions

generate_boxes

Generates box patterns for all fields.

generate_diagonals

Generates diagonal patterns for all fields.

generate_files

Generates file patterns for all fields.

generate_fronts

Generates front patterns for all fields.

generate_jumps

Generates jump patterns for all fiellds.

generate_rails

Generates rail patterns for all fields.

generate_ranks

Generates rank patterns for all fields.

generate_stars

Generates star patterns for all fields.

get_box

Get a box pattern for the field specified by field_index.

get_diagonals

Gets a diagonal pattern for the field specified by field_index.

get_file

Gets a file pattern for the field specified by field_index.

get_front

Get a front pattern for the field specified by field_index, from the color perspective.

get_jumps

Get a jumps pattern for the field specified by field_index.

get_rail

Get a rail pattern for the field specified by field_index.

get_rank

Gets a rank pattern for the field specified by field_index.

get_star

Get a star pattern for the field specified by field_index.

init

Initializes all patterns by filling arrays.