Module inanis::state::text::fen

Structs

ParsedEPD

Functions

active_color_to_fen 🔒

Converts active color from the board into the FEN chunk.

board_to_epd

Converts Board into the EPD.

board_to_fen

Converts Board into the FEN.

castling_to_fen 🔒

Converts castling rights from the board into the FEN chunk.

en_passant_to_fen 🔒

Converts en passant from the board into the FEN chunk.

epd_to_board

Converts epd into the Board, using provided containers. If the parameter is None, then the new container is created. Returns Err with proper error message if epd couldn’t be parsed correctly.

fen_to_active_color 🔒

Parses FEN’s active color and stores it into the board. Returns Err with the proper error message if active_color couldn’t be parsed.

fen_to_board

Converts fen into the Board, using provided containers. If the parameter is None, then the new container is created. Returns Err with proper error message if fen couldn’t be parsed correctly.

fen_to_castling 🔒

Parses FEN’s castling rights and stores them into the board. Returns Err with the proper error message if castling couldn’t be parsed.

fen_to_en_passant 🔒

Parses FEN’s en passant and stores it into the board. Returns Err with the proper error message if en_passant couldn’t be parsed.

fen_to_fullmove_number 🔒

Parses FEN’s fullmove number and stores it into the board. Returns Err with the proper error message if fullmove_number couldn’t be parsed.

fen_to_halfmove_clock 🔒

Parses FEN’s halfmove clock and stores it into the board. Returns Err with the proper error message if halfmove_clock couldn’t be parsed.

fen_to_pieces 🔒

Parses FEN’s pieces and stores them into the board. Returns Err with the proper error message if pieces couldn’t be parsed.

fullmove_number_to_fen 🔒

Converts fullmove number from the board into the FEN chunk.

get_epd_parameter 🔒

Gets a value of the name parameters from the specified epd. Returns None if the parameter was not found.

halfmove_clock_to_fen 🔒

Converts halfmove clock from the board into the FEN chunk.

pieces_to_fen 🔒

Converts pieces from the board into the FEN chunk.