Module inanis::tablebases::syzygy::probe

Functions

get_max_pieces_count

Gets maximal pieces count supported by loaded Syzygy tablebases. Returns 0 if the feature is disabled.

get_root_wdl_dtz

Gets WDL (Win-Draw-Loss), DTZ (Distance To Zeroing) and the best move for the position specified in board. Returns None if data couldn’t be obtained or the feature is disabled.

get_wdl

Gets WDL (Win-Draw-Loss) for the position specified in board. Returns None if data couldn’t be obtained or the feature is disabled.

init

Loads Syzygy tablebases stored in syzygy_path location.